[minor] rename webpack -> webpackConfig
