+++ packages/gasket-plugin-analyze/lib/webpack-config.js
