+++ packages/gasket-plugin-service-worker/lib/index.js
@@ -2,7 +2,7 @@ const configure = require('./configure');
 const build = require('./build');
 const middleware = require('./middleware');
 const express = require('./express');
-const webpackConfig = require('./webpack');
+const webpackConfig = require('./webpack-config');
 
 module.exports = {
   name: require('../package').name,
