+++ packages/gasket-plugin-service-worker/lib/webpack-config.js
